requirements.txt
pyyaml>=6.0.1
transformers==4.37.2
accelerate>=0.30.0
decord==0.6.0
opencv-python==4.10.0.84
av>=11.0.0
einops>=0.7.0
timm>=0.9.16
sentencepiece
easydict
huggingface_hub

tqdm
yacs
tensorboard
pytorch-pretrained-bert
Pillow
opencv_python
ffmpeg_python
scipy
cython
packaging==21.3
ftfy
matplotlib
ipykernel
jupyter
notebook
numpy


fastapi
uvicorn[standard]
pydantic
pillow